+++ snips-nlu-ontology-parsers/src/test_utils.rs
@@ -1,7 +1,0 @@
-use std::path::{PathBuf, Path};
-
-pub fn test_path() -> PathBuf {
-    Path::new("..")
-        .join("data")
-        .join("tests")
-}